Initial commit of Redmine Trac Formatter plugin
This is heavily based on https://github.com/yugui/redmine_rd_formatter,
which was the first plugin for formatting markup in Redmine.  Yugui
created the patch that made formatting plugins possible and released
the above-mentioned formatter as a template for others to follow (see
http://www.redmine.org/issues/2025#note-4)

The initial checkin is largely based on the files Yugui includes in his
project but modified to work with this one.  There are two files listed
in the TODO file that have not yet been recreated in my project.  None
of these files have been tested with the exception of
lib/redmin_trac_formatter/wiki_formatter.rb.  It has some of the basic
Trac wiki formatting complete and placeholders listed for most of the
rest of the work.

This work is released under the Apache Software License, version 2.0.
See LICENSE for more information.