add support for images and fix bug in bullet lists
If the first line of a bulleted list was indented, it created the next
li (which was indented at the same level) to be a sub-list of the first
item.  This is because the line was stripped/trimmed, which led to the
first item appearing as though it was not indented.  Now we simply chomp
the line rather than stripping it.

Added basic support for images.